style: Simple style modifications (offset, paragraph, etc.)
