+++ src/main/java/com/mcc/outify/security/WebSecurityConfig.java
@@ -1,0 +1,78 @@
+package com.mcc.outify.security;
+
+import com.mcc.outify.security.oauth.CustomOAuth2UserService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.Customizer;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.CsrfConfigurer;
+import org.springframework.security.config.annotation.web.configurers.HttpBasicConfigurer;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.CorsConfigurationSource;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+
+import java.util.Arrays;
+
+@EnableWebSecurity
+@RequiredArgsConstructor
+@Configuration
+public class WebSecurityConfig {
+
+    private final MyAuthenticationSuccessHandler oAuth2LoginSuccessHandler;
+    private final CustomOAuth2UserService customOAuth2UserService;
+    private final JwtAuthFilter jwtAuthFilter;
+    private final MyAuthenticationFailureHandler oAuth2LoginFailureHandler;
+
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+        http
+                .httpBasic(HttpBasicConfigurer::disable)
+                .cors(Customizer.withDefaults())
+                .csrf(CsrfConfigurer::disable)
+
+                // Spring Security에서 Session 생성 또는 사용 X
+                .sessionManagement(sessionManagement -> sessionManagement.sessionCreationPolicy(
+                        SessionCreationPolicy.STATELESS
+                ))
+
+                .authorizeHttpRequests(authorize -> authorize
+                        .requestMatchers("/token/**").permitAll()
+                        .requestMatchers("/", "/healthcheck", "/weathers").permitAll()
+                        // 인가 없이 접속할 주소 (** : 모든 주소) 는 권한 없이 접속 가능
+                        // 예시) .requestMatchers("/admin/**").hasRole("ADMIN")
+                        .anyRequest().authenticated() // anyRequest() : 나머지 모든 주소 (제일 마지막에 위치해야 함) 는 로그인이 필요
+                )
+
+                .oauth2Login((oauth2) -> oauth2
+                .loginPage("/oauth2/authorization/google") // 권한 접근 실패 시 로그인 페이지로 이동
+                .defaultSuccessUrl("https://outify.co.kr") // 로그인 성공 시 이동할 페이지
+                .failureUrl("/oauth2/authorization/google") // 로그인 실패 시 이동 페이지
+                .userInfoEndpoint(userInfoEndpoint -> userInfoEndpoint
+                        .userService(customOAuth2UserService))); // OAuth2 로그인시 사용자 정보를 가져오는 엔드포인트와 사용자 서비스를 설정
+
+        return http.addFilterBefore(jwtAuthFilter, UsernamePasswordAuthenticationFilter.class)
+                .build();
+    }
+
+    @Bean
+    public CorsConfigurationSource corsConfigurationSource() {
+        CorsConfiguration config = new CorsConfiguration();
+
+        // cors 설정
+        config.setAllowCredentials(true); // 실제 응답을 보낼 때, 브라우저에게 자격 증명과 함께 요청을 보낼 수 있도록 허용합니다.
+        config.setAllowedOriginPatterns(Arrays.asList("*")); // 모든 원본에서의 요청을 허용합니다.
+        config.setAllowedMethods(Arrays.asList("HEAD", "POST", "GET", "DELETE", "PUT", "PATCH")); // 허용할 HTTP 메서드를 설정합니다.
+        config.setAllowedHeaders(Arrays.asList("*")); // 모든 헤더의 요청을 허용합니다.
+
+        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+        source.registerCorsConfiguration("/**", config); // 모든 경로에 대해 위에서 설정한 CORS 설정을 적용합니다.
+
+        return source;
+    }
+
+}
