+++ src/main/java/com/mcc/outify/weathers/WeatherRequestDto.java
@@ -8,4 +8,5 @@ public class WeatherRequestDto {
 
     @NotBlank
     private String addr;
+
 }
