+++ src/main/java/com/mcc/outify/weathers/WeatherResponseDto.java
@@ -34,4 +34,5 @@ public class WeatherResponseDto {
         this. hum = weatherData.getHum();
         this.dpt = weatherData.getDpt();
     }
+
 }
