+++ src/main/java/com/mcc/outify/weathers/entity/WeatherDataEntity.java
@@ -75,4 +75,5 @@ public class WeatherDataEntity {
         this.hum = weatherData.getHum();
         this.dpt = weatherData.getDpt();
     }
+
 }
