+++ src/main/java/com/mcc/outify/weathers/openApi/YrAPI.java
@@ -109,4 +109,5 @@ public class YrAPI implements WeatherAPI {
             }
         }
     }
+
 }
